refactor(wallets): remove the legacy provider runtime
Every provider registers only a hub version, so the runtime that drove legacy
ones was dead code. `useProviders` collapses onto the hub adapter, and the
branches, storage and helpers that only ran for legacy providers go with it.

The widget's legacy provider type is deprecated rather than removed, so
integrators keep compiling while they migrate. On the react side only the
wallet-facing types are left in `legacy/`; the context, hooks and auto-connect
glue moved next to what uses them.
