--- wallets/react/helpers/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "@rango-dev/wallets-react/helpers",
-  "type": "module",
-  "main": "../dist/helpers/index.js",
-  "module": "../dist/helpers/index.js",
-  "types": "../dist/helpers/index.d.ts",
-  "sideEffects": false
-}
