--- wallets/react/src/helpers/index.ts
@@ -1,1 +1,0 @@
-export { getAllLegacyProviders } from '../hub/mod.js';
