+++ wallets/react/src/hub/mod.ts
@@ -1,6 +1,2 @@
-export {
-  separateLegacyAndHubProviders,
-  getAllLegacyProviders,
-  findProviderByType,
-} from './utils.js';
+export { getHubProviders, findProviderByType } from './utils.js';
 export { useHubAdapter } from './useHubAdapter.js';
