+++ wallets/react/src/hub/useHubAdapter.ts
@@ -1,7 +1,7 @@
 import type { AllProxiedNamespaces, ExtensionLink } from './types.js';
-import type { ProviderContext } from '../index.js';
+import type { ConnectResult } from '../legacy/mod.js';
+import type { ProviderContext, ProviderProps } from '../types.js';
 import type { Provider } from '@hub3js/core';
-import type { VersionedProviders } from '@hub3js/core/utils';
 import type { Accounts, AccountsWithActiveChain } from '@hub3js/std/types';
 import type {
   LegacyNamespaceInputForConnect,
@@ -17,18 +17,16 @@ import {
 import { useEffect, useRef, useState } from 'react';
 import { Ok, Result } from 'ts-results';
 
-import {
-  type ConnectResult,
-  HUB_LAST_CONNECTED_WALLETS,
-  type ProviderProps,
-} from '../legacy/mod.js';
-import { useAutoConnect } from '../legacy/useAutoConnect.js';
+import { withErrorLoggingApi } from '../helpers.js';
 
 import { autoConnect } from './autoConnect.js';
+import { HUB_LAST_CONNECTED_WALLETS } from './constants.js';
 import { createQueue, fromAccountIdToLegacyAddressFormat } from './helpers.js';
 import { LastConnectedWalletsFromStorage } from './lastConnectedWallets.js';
+import { useAutoConnect } from './useAutoConnect.js';
 import { useHubRefs } from './useHubRefs.js';
 import {
+  findProviderByType,
   isEvmNamespace,
   isSolanaNamespace,
   isUtxoNamespace,
@@ -39,8 +37,6 @@ import {
 
 export type UseAdapterParams = Omit<ProviderProps, 'providers'> & {
   providers: Provider[];
-  /** This is only will be used to access some parts of the legacy provider that doesn't exists in Hub. */
-  allVersionedProviders: VersionedProviders[];
 };
 
 export function useHubAdapter(params: UseAdapterParams): ProviderContext {
@@ -49,7 +45,6 @@ export function useHubAdapter(params: UseAdapterParams): ProviderContext {
   // useEffect will run `subscribe` once, so we need a reference and mutate the value if it's changes.
   const dataRef = useRef({
     onUpdateState: params.onUpdateState,
-    allVersionedProviders: params.allVersionedProviders,
     allBlockChains: params.allBlockChains,
   });
   const hubInitiated = useRef(false);
@@ -80,7 +75,6 @@ export function useHubAdapter(params: UseAdapterParams): ProviderContext {
   useEffect(() => {
     dataRef.current = {
       onUpdateState: params.onUpdateState,
-      allVersionedProviders: params.allVersionedProviders,
       allBlockChains: params.allBlockChains,
     };
   }, [params]);
@@ -140,9 +134,8 @@ export function useHubAdapter(params: UseAdapterParams): ProviderContext {
       const provider = getHub().get(type);
 
       if (!provider) {
-        throw new Error(
-          `You should add ${type} to provider first then call 'canSwitchNetworkTo'.`
-        );
+        // TODO: should throw an error
+        return false;
       }
 
       if (!namespace) {
@@ -436,18 +429,22 @@ export function useHubAdapter(params: UseAdapterParams): ProviderContext {
         properties: metadata.properties,
       };
     },
-    providers() {
-      throw new Error('This method is not available on hub providers.');
-    },
     state(type) {
       const hubState = getHub().state();
       const wallet = getHub().get(type);
       const walletState = hubState[type];
 
       if (!walletState || !wallet) {
-        throw new Error(
-          `It seems your requested provider doesn't exist in hub. Provider Id: ${type}`
-        );
+        // TODO: should throw an error
+        return {
+          connected: false,
+          connecting: false,
+          installed: false,
+          reachable: false,
+          accounts: null,
+          network: null,
+          namespaces: new Map(),
+        };
       }
 
       const accounts = walletState.namespaces
@@ -501,12 +498,18 @@ export function useHubAdapter(params: UseAdapterParams): ProviderContext {
         `Suggest has not been implemented for given chain: ${namespace.network}`
       );
     },
-    hubProvider() {
-      throw new Error(
-        'Unreachable code. the method has been implemented in main adapter instance.'
-      );
+    hubProvider(type) {
+      const provider = findProviderByType(params.providers, type);
+
+      if (!provider) {
+        throw new Error(
+          `You're trying to access ${type} provider which is not found in hub providers. it may not be registered yet.`
+        );
+      }
+
+      return provider;
     },
   };
 
-  return api;
+  return withErrorLoggingApi(api);
 }
