+++ wallets/react/src/hub/utils.ts
@@ -1,4 +1,6 @@
 import type { AllProxiedNamespaces } from './types.js';
+import type { ConnectResult } from '../legacy/mod.js';
+import type { ProviderProps } from '../types.js';
 import type { UtxoActions } from '@hub3js/bip122';
 import type { Hub, Provider, ProxiedNamespace } from '@hub3js/core';
 import type { Event } from '@hub3js/core/store';
@@ -9,7 +11,6 @@ import type {
 import type { SolanaActions } from '@hub3js/solana';
 import type {
   LegacyNamespaceInputForConnect,
-  LegacyProviderInterface,
   LegacyEventHandler as WalletEventHandler,
   LegacyWalletType as WalletType,
 } from '@rango-dev/wallets-core/legacy';
@@ -27,12 +28,7 @@ import {
 import { AccountId } from 'caip';
 import { type BlockchainMeta, isEvmBlockchain } from 'rango-types';
 
-import {
-  type ConnectResult,
-  HUB_LAST_CONNECTED_WALLETS,
-  type ProviderProps,
-} from '../legacy/mod.js';
-
+import { HUB_LAST_CONNECTED_WALLETS } from './constants.js';
 import {
   fromAccountIdToLegacyAddressFormat,
   isConnectResultEvm,
@@ -40,27 +36,16 @@ import {
 } from './helpers.js';
 import { LastConnectedWalletsFromStorage } from './lastConnectedWallets.js';
 
-/* Gets a list of hub and legacy providers and returns a tuple which separates them. */
-export function separateLegacyAndHubProviders(
-  providers: VersionedProviders[]
-): [LegacyProviderInterface[], Provider[]] {
-  const LEGACY_VERSION = '0.0.0';
-  const HUB_VERSION = '1.0.0';
-
-  const legacyProviders: LegacyProviderInterface[] = [];
-  const hubProviders: Provider[] = [];
-
-  providers.forEach((provider) => {
+export function getHubProviders(providers: VersionedProviders[]): Provider[] {
+  return providers.map((provider) => {
     try {
-      const target = pickVersion(provider, HUB_VERSION);
-      hubProviders.push(target[1]);
+      return pickVersion(provider, '1.0.0')[1] as Provider;
     } catch {
-      const target = pickVersion(provider, LEGACY_VERSION);
-      legacyProviders.push(target[1]);
+      throw new Error(
+        "Legacy providers aren't supported anymore. Expected a provider with version '1.0.0'."
+      );
     }
   });
-
-  return [legacyProviders, hubProviders];
 }
 
 export function findProviderByType(
@@ -270,21 +255,6 @@ export function mapHubEventsToLegacy(
       });
       break;
   }
-}
-
-export function getAllLegacyProviders(
-  allProviders: VersionedProviders[]
-): LegacyProviderInterface[] {
-  const LEGACY_VERSION = '0.0.0';
-
-  const legacyProviders: LegacyProviderInterface[] = [];
-
-  allProviders.forEach((provider) => {
-    const target = pickVersion(provider, LEGACY_VERSION);
-    legacyProviders.push(target[1]);
-  });
-
-  return legacyProviders;
 }
 
 /**
