+++ wallets/react/src/legacy/mod.ts
@@ -1,13 +1,1 @@
-export type {
-  ProviderProps,
-  ProviderContext,
-  ConnectResult,
-  ExtendedWalletInfo,
-} from './types.js';
-export {
-  LEGACY_LAST_CONNECTED_WALLETS,
-  HUB_LAST_CONNECTED_WALLETS,
-} from '../hub/constants.js';
-
-export { WalletContext } from './context.js';
-export { useLegacyProviders } from './useLegacyProviders.js';
+export type { ConnectResult, ExtendedWalletInfo } from './types.js';
