--- wallets/react/src/test-utils/fixtures.ts
@@ -1,164 +1,0 @@
-import type { LegacyProviderInterface } from '@rango-dev/wallets-core/legacy';
-import type { BlockchainMeta, GenericSigner } from 'rango-types';
-import type { EvmTransaction } from 'rango-types/mainApi';
-
-import { DefaultSignerFactory, TransactionType } from 'rango-types';
-
-export const legacyAddress = '0x000000000000000000000000000000000000dead';
-
-export class MockEvmSigner implements GenericSigner<EvmTransaction> {
-  async signMessage(
-    _msg: string,
-    _address: string,
-    _chainId: string | null
-  ): Promise<string> {
-    return '0x';
-  }
-  async signAndSendTx(
-    _tx: EvmTransaction,
-    _address: string,
-    _chainId: string | null
-  ): Promise<{ hash: string; response?: unknown }> {
-    return {
-      hash: '0x',
-    };
-  }
-}
-
-export const legacyProvider: LegacyProviderInterface = {
-  config: {
-    type: 'legacy-garbage',
-  },
-  async connect({ network }) {
-    // we added a timeout to simulate requesting to extension for connection.
-    return await new Promise((resolve, reject) => {
-      setTimeout(() => {
-        // This is useful for simulating error scenarios.
-        if (network === 'When Airdrop?') {
-          reject(
-            'please stay tuned we shall announce the detailed information soon.'
-          );
-        } else {
-          resolve([
-            {
-              accounts: [legacyAddress],
-              chainId: network || '',
-            },
-          ]);
-        }
-        // eslint-disable-next-line @typescript-eslint/no-magic-numbers
-      }, 10);
-    });
-  },
-  getInstance() {
-    return {};
-  },
-  async getSigners() {
-    const signers = new DefaultSignerFactory();
-    signers.registerSigner(TransactionType.EVM, new MockEvmSigner());
-    return signers;
-  },
-  getWalletInfo() {
-    return {
-      name: 'legacy garbage wallet',
-      color: '#000',
-      img: 'https://...',
-      installLink: 'https://...',
-      supportedChains: [],
-    };
-  },
-};
-
-export const blockchainsMeta: BlockchainMeta[] = [
-  {
-    name: 'BTC',
-    defaultDecimals: 8,
-    addressPatterns: [
-      '^[13][a-km-zA-HJ-NP-Z1-9]{25,34}$|^(bc1)[0-9A-Za-z]{39,59}$',
-    ],
-    feeAssets: [
-      {
-        blockchain: 'BTC',
-        symbol: 'BTC',
-        address: null,
-      },
-    ],
-    logo: 'https://raw.githubusercontent.com/rango-exchange/assets/main/blockchains/BTC/icon.svg',
-    displayName: 'Bitcoin',
-    shortName: 'BTC',
-    sort: 12,
-    color: '#F7931A',
-    enabled: true,
-    type: TransactionType.TRANSFER,
-    chainId: null,
-    info: {
-      infoType: 'TransferMetaInfo',
-      blockExplorerUrls: ['https://www.blockchain.com/btc/'],
-      addressUrl: 'https://www.blockchain.com/btc/address/{wallet}',
-      transactionUrl: 'https://www.blockchain.com/btc/tx/{txHash}',
-      tokenUrl: null,
-    },
-  },
-  {
-    name: 'ETH',
-    defaultDecimals: 18,
-    addressPatterns: ['^(0x)[0-9A-Fa-f]{40}$'],
-    feeAssets: [
-      {
-        blockchain: 'ETH',
-        symbol: 'ETH',
-        address: null,
-      },
-    ],
-    logo: 'https://raw.githubusercontent.com/rango-exchange/assets/main/blockchains/ETH/icon.svg',
-    displayName: 'Ethereum',
-    shortName: 'ETH',
-    sort: 0,
-    color: '#ecf0f1',
-    enabled: true,
-    type: TransactionType.EVM,
-    chainId: '0x1',
-    info: {
-      infoType: 'EvmMetaInfo',
-      chainName: 'Ethereum Mainnet',
-      nativeCurrency: {
-        name: 'ETH',
-        symbol: 'ETH',
-        decimals: 18,
-      },
-      rpcUrls: ['https://rpc.ankr.com/eth'],
-      blockExplorerUrls: ['https://etherscan.io'],
-      addressUrl: 'https://etherscan.io/address/{wallet}',
-      transactionUrl: 'https://etherscan.io/tx/{txHash}',
-      enableGasV2: true,
-      tokenUrl: 'https://etherscan.io/token/{address}',
-    },
-  },
-  {
-    name: 'SOLANA',
-    defaultDecimals: 9,
-    addressPatterns: ['^[1-9A-HJ-NP-Za-km-z]{32,44}$'],
-    feeAssets: [
-      {
-        blockchain: 'SOLANA',
-        symbol: 'SOL',
-        address: null,
-      },
-    ],
-    logo: 'https://raw.githubusercontent.com/rango-exchange/assets/main/blockchains/SOLANA/icon.svg',
-    displayName: 'Solana',
-    shortName: 'Solana',
-    sort: 20,
-    color: '#708DD2',
-    enabled: true,
-    type: TransactionType.SOLANA,
-    chainId: 'mainnet-beta',
-    info: {
-      infoType: 'SolanaMetaInfo',
-      blockExplorerUrls: ['https://solscan.io/'],
-      addressUrl: 'https://solscan.io/account/{wallet}',
-      transactionUrl: 'https://solscan.io/tx/{txHash}',
-      tokenUrl: 'https://solscan.io/token/{address}',
-    },
-  },
-];
