+++ widget/embedded/src/types/config.ts
@@ -278,6 +278,10 @@ export type WidgetConfig = {
   from?: BlockchainAndTokenConfig;
   to?: BlockchainAndTokenConfig;
   liquiditySources?: string[];
+  /**
+   * Passing a `LegacyProviderInterface` is deprecated. Pass a hub `Provider` instead.
+   * Legacy providers will be removed in future versions.
+   */
   wallets?: (WalletType | LegacyProviderInterface | Provider)[];
   multiWallets?: boolean;
   customDestination?: boolean;
