+++ widget/embedded/src/utils/providers.ts
@@ -1,5 +1,4 @@
 import type { WidgetConfig } from '../types';
-import type { LegacyProviderInterface } from '@rango-dev/wallets-core/legacy';
 
 import { Provider } from '@hub3js/core';
 import {
@@ -19,11 +18,10 @@ export interface ProvidersOptions {
 
 /**
  *
- * Generate a list of providers by passing a provider name (e.g. metamask) or a custom provider which implemented ProviderInterface.
- * @returns BothProvidersInterface[] a list of BothProvidersInterface
+ * Generate a list of providers by passing a provider name (e.g. metamask) or a custom hub `Provider`.
+ * @returns VersionedProviders[] a list of VersionedProviders
  *
  */
-type BothProvidersInterface = LegacyProviderInterface | Provider;
 export function matchAndGenerateProviders({
   allProviders,
   configWallets,
@@ -43,30 +41,13 @@ export function matchAndGenerateProviders({
       /*
        * There are two types of provider we get, the first one is only passing the wallet name
        * then we will match the wallet name with our providers (@rango-dev/provider-*).
-       * The second way is passing a custom provider which implemented ProviderInterface.
+       * The second way is passing a custom hub `Provider` instance.
        */
       if (typeof requestedWallet === 'string') {
-        const result = allProviders.find((provider) => {
-          /*
-           * To find a provider in allProviders,
-           * a version of each provider should be picked
-           * and validated based on that version scheme.
-           * If the corresponding provider to a wallet was found in allProvider,
-           * it will be add to selected providers.
-           */
-          const versionedProvider =
-            pickProviderVersionWithFallbackToLegacy(provider);
-          if (versionedProvider instanceof Provider) {
-            return versionedProvider.id === requestedWallet;
-          }
-          return versionedProvider.config.type === requestedWallet;
-        });
+        const result = allProviders.find(
+          (provider) => pickProviderVersion(provider).id === requestedWallet
+        );
 
-        /*
-         * A provider may have multiple versions
-         * (e.g., 0.0, also known as legacy, and 1.0, also known as hub).
-         * We should ensure that all existing versions of a provider are added to selectedProviders.
-         */
         if (result) {
           selectedProviders.push(result);
         } else {
@@ -77,15 +58,15 @@ export function matchAndGenerateProviders({
         }
       } else {
         // It's a custom provider so we directly push it to the list.
-        if (requestedWallet instanceof Provider) {
-          selectedProviders.push(
-            defineVersions().version('1.0.0', requestedWallet).build()
-          );
-        } else {
-          selectedProviders.push(
-            defineVersions().version('0.0.0', requestedWallet).build()
+        if (!(requestedWallet instanceof Provider)) {
+          throw new Error(
+            `Legacy providers aren't supported anymore. Pass a hub 'Provider' instance in 'wallets' instead.`
           );
         }
+
+        selectedProviders.push(
+          defineVersions().version('1.0.0', requestedWallet).build()
+        );
       }
     });
 
@@ -95,27 +76,12 @@ export function matchAndGenerateProviders({
   return allProviders;
 }
 
-export function pickProviderVersionWithFallbackToLegacy(
-  provider: VersionedProviders
-): BothProvidersInterface {
-  try {
-    return pickVersion(provider, '1.0.0')[1];
-  } catch {
-    // Fallback to legacy version, if target version doesn't exists.
-    return pickVersion(provider, '0.0.0')[1];
-  }
+export function pickProviderVersion(provider: VersionedProviders): Provider {
+  return pickVersion(provider, '1.0.0')[1];
 }
 
 export function configWalletsToWalletName(
   providers: VersionedProviders[]
 ): string[] {
-  const names = providers
-    .map((provider) => pickProviderVersionWithFallbackToLegacy(provider))
-    .map((provider) => {
-      if (provider instanceof Provider) {
-        return provider.id;
-      }
-      return provider.config.type;
-    });
-  return names;
+  return providers.map((provider) => pickProviderVersion(provider).id);
 }
