+++ widget/playground/src/containers/FunctionalLayout/FunctionalLayout.utils.ts
@@ -1,6 +1,5 @@
 import type { MapSupportedList } from '../../components/MultiSelect/MultiSelect.types';
 import type { Provider } from '@hub3js/core';
-import type { ProviderInterface } from '@rango-dev/wallets-react';
 import type { WidgetConfig } from '@rango-dev/widget-embedded';
 import type { BlockchainMeta } from 'rango-sdk';
 
@@ -29,47 +28,30 @@ export function getWalletsList(
   const allBuiltProviders = allProviders.map((build) => build());
   const walletsList: MapSupportedList[] = [];
   allBuiltProviders.forEach((versionedProvider: VersionedProviders) => {
-    try {
-      /*
-       * `pickVersion` can't narrow the value type here, because `VersionedProviders`
-       * doesn't carry the version literals of a specific provider.
-       */
-      const provider = pickVersion(versionedProvider, '1.0.0')[1] as Provider;
-      if (excludedWallets.includes(provider.id)) {
-        return;
-      }
-      const info = provider.info();
-      if (!info) {
-        throw new Error('Provider info is not available.');
-      }
-
-      const supportedChains = getSupportedChainsFromProvider(
-        provider,
-        blockchains
-      );
-      walletsList.push({
-        title: info.metadata.name,
-        logo: info.metadata.icon,
-        name: provider.id,
-        supportedNetworks: getCategoryNetworks(supportedChains),
-      });
-    } catch {
-      // Fallback to legacy version, if target version doesn't exists.
-      const provider = pickVersion(
-        versionedProvider,
-        '0.0.0'
-      )[1] as ProviderInterface;
-      if (excludedWallets.includes(provider.config.type)) {
-        return;
-      }
-      const walletInfo = provider.getWalletInfo(blockchains);
-      walletsList.push({
-        title: walletInfo.name,
-        logo: walletInfo.img,
-        name: provider.config.type,
-        supportedNetworks: getCategoryNetworks(walletInfo.supportedChains),
-      });
+    /*
+     * `pickVersion` can't narrow the value type here, because `VersionedProviders`
+     * doesn't carry the version literals of a specific provider.
+     */
+    const provider = pickVersion(versionedProvider, '1.0.0')[1] as Provider;
+    if (excludedWallets.includes(provider.id)) {
+      return;
     }
+
+    const info = provider.info();
+    if (!info) {
+      throw new Error('Provider info is not available.');
+    }
+
+    const supportedChains = getSupportedChainsFromProvider(
+      provider,
+      blockchains
+    );
+    walletsList.push({
+      title: info.metadata.name,
+      logo: info.metadata.icon,
+      name: provider.id,
+      supportedNetworks: getCategoryNetworks(supportedChains),
+    });
   });
 
   return walletsList;
